reorganized the application structure and added new api router to the application for language operations. Followed the documentation - https://fastapi.tiangolo.com/tutorial/bigger-applications/
