+++ app/main.py
@@ -1,6 +1,7 @@
-from fastapi import FastAPI, HTTPException, Response
-from .userprofile.user_service import UserService
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
+from .userprofile.user_info import user_info_router
+from .userprofile.language_operations import language_operations_router
 
 tags_metadata = [
     {
@@ -10,6 +11,10 @@ tags_metadata = [
     {
         "name": "Github User Operations",
         "description": "These operations fetches various information of the user's github profile."
+    },
+    {
+        "name": "Languages used operations",
+        "description": "These are responsible for computing language statistics based on users github profile."
     }
 ]
 
@@ -21,6 +26,9 @@ app = FastAPI(
     docs_url="/swagger",
     redoc_url="/"
 )
+
+app.include_router(user_info_router)
+app.include_router(language_operations_router)
 
 origins = [
     "http://localhost",
@@ -40,99 +48,3 @@ app.add_middleware(
 @app.get("/", tags=["Fast API demo"])
 def read_root():
     return "This API is built using FAST API , which performs some operations using Github API"
-
-
-@app.get("/getUserDetails/{username}", tags=["Github User Operations"])
-def getUserDetails(username: str):
-    try:
-        userService = UserService(username)
-        userDetails = userService.getUserDetails()
-        print(userDetails)
-        return userDetails
-    except Exception as e:
-        raise HTTPException(
-            status_code=404,
-            detail=e,
-        )
-
-
-@app.get("/getUserFollowers/{username}", tags=["Github User Operations"])
-def getUserFollowers(username: str):
-    try:
-        userService = UserService(username)
-        followers = userService.getUserFollowers()
-        return followers
-    except Exception as e:
-        raise HTTPException(
-            status_code=404,
-            detail=e
-        )
-
-
-@app.get("/getUserGists/{username}", tags=["Github User Operations"])
-def getUserGists(username: str):
-    try:
-        userService = UserService(username)
-        if userService.isValidGithubUser():
-            gists = userService.getUserGists()
-            return gists
-        else:
-            print("user not found")
-            return Response(status_code=404, content="Github user not found")
-
-    except Exception as e:
-        raise HTTPException(
-            status_code=404,
-            detail=e
-        )
-
-
-@app.get("/getUserGist/{username}/{gist_id}", tags=["Github User Operations"])
-def getUserGist(username: str, gist_id: str):
-    try:
-        userService = UserService(username)
-
-        if userService.isValidGithubUser():
-            gist = userService.getUserGist(gist_id)
-            return gist
-        else:
-            return Response(status_code=404, content="Github user not found")
-    except Exception as e:
-        raise HTTPException(
-            status_code=500,
-            detail=e
-        )
-
-
-@app.get("/getUserRepos/{username}", tags=["Github User Operations"])
-def getUserRepos(username: str):
-    try:
-        userService = UserService(username)
-
-        if userService.isValidGithubUser():
-            repos = userService.getUserRepos()
-            return repos
-        else:
-            return Response(status_code=404, content="Github user not found")
-    except Exception as e:
-        raise HTTPException(
-            status_code=500,
-            detail=e
-        )
-
-
-@app.get("getUserRepoLanguages/{username}/{repo_name}", tags=["Github User Operations"])
-def getUserRepoLanguages(username: str, repo_name: str):
-    try:
-        userService = UserService(username)
-
-        if userService.isValidGithubUser():
-            languages = userService.getUserRepoLanguages(repo_name)
-            return languages
-        else:
-            return Response(status_code=404, content="Github user not found")
-    except Exception as e:
-        raise HTTPException(
-            status_code=500,
-            detail=e
-        )
