+++ app/userprofile/language_service.py
@@ -1,4 +1,3 @@
-import requests
 from typing import Dict
